Add enriched collections cache collection

+++ public/js/index.js
@@ -9,9 +9,10 @@ $("#submit-form").on("click", async (e) => {
 
   // Check if user exists
   const getCollectionUrl = `${API_URL}/collections/${username}`;
-  const data = await axios.get(getCollectionUrl, { validateStatus: false });
-  const status = data.status;
+  let data = await axios.get(getCollectionUrl, { validateStatus: false });
+  let status = data.status;
 
+  // If user exists, proceed
   if (status == 200 || status == 201) {
     $(".invalid-feedback").remove();
     $("#login-form-text").removeClass("is-invalid");
@@ -26,34 +27,43 @@ $("#submit-form").on("click", async (e) => {
         <span aria-hidden="true">&times;</span>
       </button>
     </div>`);
-    }, 30000);
-
+    }, 45000);
     $(".alert-warning").css("border-width", "0px");
-
     // Add Loading Screen
     $("#submit-form").text("")
       .append(`<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
     Processing Collection...`);
 
-    // Read Collection
-    // const getCollectionUrl = `${API_URL}/collections/${username}`;
-    const getEnrichedCollectionUrl = `${API_URL}/collections/${username}/enrich`;
-    // await axios(getCollectionUrl);
-    const response = await axios(getEnrichedCollectionUrl);
-    const collection = response.data;
-    const compressedCollection = LZUTF8.compress(JSON.stringify(collection), {
-      outputEncoding: "StorageBinaryString",
-    });
+    // Rnrich collection
+    await axios.post(`${API_URL}/collections/${username}/enrich`);
 
-    
-    $(".spinner-border").remove();
-    $("#submit-form").text("All Done!").css("background-color", LIGHT_COLOR);
-    window.localStorage.clear();
-    window.localStorage.setItem("collection", compressedCollection);
+    // Periodically check if collection was enriched
+    const intervalID = setInterval(async function () {
+      const response = await axios.get(
+        `${API_URL}/collections/${username}/enrich`
+      );
+      if (response.status == 200) {
+        await clearInterval(intervalID);
+        const collection = response.data;
+        const compressedCollection = LZUTF8.compress(
+          JSON.stringify(collection),
+          {
+            outputEncoding: "StorageBinaryString",
+          }
+        );
 
-    await setTimeout((_) => {
-      window.open(`${API_URL}/main`, "_self");
-    }, 700);
+        $(".spinner-border").remove();
+        $("#submit-form")
+          .text("All Done!")
+          .css("background-color", LIGHT_COLOR);
+        window.localStorage.clear();
+        window.localStorage.setItem("collection", compressedCollection);
+
+        await setTimeout((_) => {
+          window.open(`${API_URL}/main`, "_self");
+        }, 700);
+      }
+    }, 5000);
   } else if (username == "") {
     // Reset Submit Button Text
     $("#submit-form").empty().text("Let's Start!");
